# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePKL Backend API",
        "description": "Backend for ePKL mobile app",
        "contact": {
            "name": "Muhamamd Haikal",
            "email": "realityinaship@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/pkl-steps": {
            "get": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Get all Pkl Steps",
                "description": "Retrieve all Pkl Steps.",
                "operationId": "getAllPklSteps",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Get active Pkl Steps",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pkl Steps retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Create a Pkl Step",
                "description": "Create a new Pkl Step.",
                "operationId": "createPklStep",
                "requestBody": {
                    "description": "Pkl Step details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "dates": {
                                        "type": "string"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pkl Step created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Delete all Pkl Steps",
                "description": "Delete all Pkl Steps.",
                "operationId": "deleteAllPklSteps",
                "responses": {
                    "201": {
                        "description": "All Pkl Steps deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pkl-steps/active": {
            "get": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Get the active Pkl Step",
                "description": "Retrieve the active Pkl Step.",
                "operationId": "getActivePklStep",
                "responses": {
                    "200": {
                        "description": "Active Pkl Step retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pkl-steps/{id}": {
            "put": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Update a Pkl Step",
                "description": "Update an existing Pkl Step.",
                "operationId": "updatePklStep",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Pkl Step",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pkl Step details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "dates": {
                                        "type": "string"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pkl Step updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pkl Steps"
                ],
                "summary": "Delete a Pkl Step",
                "description": "Delete a Pkl Step by ID.",
                "operationId": "deletePklStep",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Pkl Step",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pkl Step deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}